+++ app/Http/Controllers/BarController.php
@@ -19,13 +19,12 @@ use Illuminate\Support\Facades\Cache;
 use Kami\RecipeUtils\UnitConverter\Units;
 use Kami\Cocktail\External\BarOptionsEnum;
 use Kami\Cocktail\Http\Requests\BarRequest;
+use Kami\Cocktail\Jobs\StartBarOptimization;
 use Kami\Cocktail\Models\Enums\UserRoleEnum;
 use Kami\Cocktail\Http\Resources\BarResource;
 use Kami\Cocktail\Models\Enums\BarStatusEnum;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Kami\Cocktail\Http\Resources\BarMembershipResource;
-use Kami\Cocktail\Jobs\StartBarOptimization;
-use Kami\Cocktail\Services\BarOptimizerService;
 
 class BarController extends Controller
 {