+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## New
 - Added `thumb_url` to `Image` schema
 - Added `image` to `CocktailBasic` and `IngredientBasic` schemas
+- Added POST `/bars/{id}/optimize` method
+    - Used to run some optional commands per bar like, recalculating ABV, refreshing bar search index, recalculating ingredient hierarchy
 
 ## Fixes
 - Fixed `Menu` and `MenuRequest` schemas being out of sync